Allow bar order to be conifgured in config.h
This patch adds char *barlayout to config(.def).h which can contain as
many or as few of the characters 'l', 'n', 's', 't', or '|' to configure
the order of dwm's bar.

    l   Layout indicator;
    n   Window name;
    s   Status (set with xsetroot);
    t   Tag indicators;
    |   Split point.

Everything before '|' (and if it is omitted) will appear on the left of
the bar. Everything after will appear on the right, but in reverse
order.